PROD-205 #time 1h Force everything to be Float objects. TODO Handle difference between occurrence and frequency with two different parameters ?

+++ trunk/ttc-term-suite/sources/eu/project/ttc/tools/indexer/IndexerView.java
@@ -292,11 +292,11 @@ public class IndexerView extends JTabbedPane implements ToolView, IndexerBinding
     }
 
     @Override
-    public void setFrequencyThreshold(Integer frequencyThreshold) {
+    public void setFrequencyThreshold(Float frequencyThreshold) {
         compEConfig.setFrequencyThreshold(frequencyThreshold);
     }
     @Override
-    public Integer getFrequencyThreshold() {
+    public Float getFrequencyThreshold() {
         return compEConfig.getFrequencyThreshold();
     }
     @Override
